Add an Class assignment (complex)
In this program, it was necessary to create a console program that meets the following requirements:
• Use the capabilities of OOP: classes, inheritance, polymorphism,
encapsulation
• Each class should have a name and informative composition to reflect
sense.
• Inheritance should only be used when it makes sense.
• The console menu should be minimal.
• Files can be used to save initialization settings. 
My task was - Sound recording. Determine the hierarchy of musical compositions. Burn to disk
collection. Calculate the duration. Rearrange the compositions of the disk with
taking into account belonging to the style. Find the song that matches
given range of track lengths.